refactor: fade functions
